Убрал геттеры и переписал все в абстрактном виде
